refactor: modularize definition functions and utilities for better organization
- Separated definition functions (e.g., `defineTask`, `defineHook`, `defineResource`, etc.) into individual files for improved maintainability and clarity.
- Re-exported all definition functions and utility type guards from a centralized location to streamline imports.
- Enhanced the structure of the codebase, making it easier to navigate and understand the relationships between different components.
- Updated references and imports throughout the code to align with the new modular structure.
